modifica parametro potenza e add database autoelettriche

+++ MODELS/automobileelettrica.h
@@ -11,7 +11,7 @@ public:
      * @param for MotoreElettrico(potenza, numeroMotori, inCarica, capacitaBatteria, capacitaAttuale, caricaSupportata, colonninaAttuale)
      */
     AutomobileElettrica(string targa, string posizione, unsigned long km,
-                        unsigned short potenza, unsigned short numeroMotori, double capacitaBatteria, double capacitaAttuale,
+                        unsigned int potenza, unsigned short numeroMotori, double capacitaBatteria, double capacitaAttuale,
                         VelocitaRicarica caricaSupportata, bool inCarica = false, Colonnina colonninaAttuale = nessuna,
                         unsigned short numeroPosti = 2, unsigned short ingombro = 3);
