+++ MODELS/motoelettrica.h
@@ -16,7 +16,7 @@ public:
      * @param ingombro
      */
     MotoElettrica(string targa, string posizione, unsigned long km, double capacitaBatteria, double capacitaAttuale,
-                  VelocitaRicarica caricaSupportata, unsigned short potenza, unsigned short numeroMotori, bool inCarica = false, Colonnina colonninaAttuale = nessuna,
+                  VelocitaRicarica caricaSupportata, unsigned int potenza, unsigned short numeroMotori, bool inCarica = false, Colonnina colonninaAttuale = nessuna,
                   unsigned short numeroPosti = 1, unsigned short ingombro = 2);
 };
